Migrate into main generate resources file
