+++ setup.py
@@ -27,7 +27,7 @@ setup(name='gilda',
       install_requires=['regex', 'adeft>=0.4.0', 'boto3', 'flask',
                         'flask-wtf', 'flask-bootstrap', 'obonet'],
       extras_require={'test': ['nose', 'coverage'],
-                      'terms': ['indra', 'tqdm'],
+                      'terms': ['indra'],
                       'benchmarks': ['pandas', 'requests']},
       keywords=['nlp', 'biology']
       )
